radar/requirements.txt
pyserial
matplotlib
numpy
python-pptx
fpdf2
